Added a format for `Some[T]`
This is to avoid confusing behaviour if you for instance try and update using `Some(x)`,  rather than `Option(x)`. We never want to fall back to an auto-derived instance in this case.

Also added the actual AttributeValue found to `NoPropertyOfType`, as that was useful when tracking down this problem.