Add ruby version
So that contributing to documentation is easier.

This is because the microsite plugin has a secret dependency on jekyll
and the latest version of jekyll is incompatible with the shipped with
of ruby on macOS. The only real competitor to rbenv is rvm and I don't
think this change is incompatible with that. With this file and rbenv
installed contributers can run gem install jekyll and things will mostly
just work.