Allow specifying port on scanamo-testkit LocalDynamoDB client
This is useful for multi-project builds that have some subprojects with
tests that require LocalDynamoDB, and others that don't - the two
subprojects will need different ports to not interfere with each other
(not killing each other's LocalDynamoDB process before time). This problem
seemed to rear it's head in the Ophan build as we tried to upgrade to
sbt 1.3.

See also https://github.com/localytics/sbt-dynamodb/issues/53 and the
alternative strategy used in Scanamo itself: https://github.com/scanamo/scanamo/pull/232