Change how SemiAuto- and AutoDerivation works
The problem was, that the derivation would (sometimes) prioritise the
derived instance instead of the one with higher priority in the
companion.

See the `AutoDerivationTest` for an example.

The fix is calling the Magnolia.gen macro directly in both derivation
traits and removing the whole `Exported` and `Hidden` logic.

See also: #583