Remove instance-method version of coercedXmap
The instance method on the `DynamoFormat` trait delegates to the static
`DynamoFormat.coercedXmap` method - but that requires an implicit
`ClassTag` for the `Throwable` type that it can coerce. The instance
method version of `coercedXmap` can't know what that `Throwable` type is,
so far as I can see, with it's current signature? The Scala 3 compiler
spots this, and errors with a `No ClassTag available` message:

```
[error] -- Error: /home/roberto/development/scanamo/scanamo/src/main/scala/org/scanamo/DynamoFormat.scala:60:58
[error] 60 |    DynamoFormat.coercedXmap(read, write)(this, implicitly)
[error]    |                                                          ^
[error]    | No ClassTag available for T
[error]    |
[error]    | where:    T is a type variable with constraint >: Null and <: Throwable
```

...the Scala 2.13 compiler doesn't seem to have a problem with it, but
probably should?!

As the `coercedXmap` instance method does not actually allow users to
specify what subtype of `Throwable` they want to coerce, I suspect it's
unused. Indeed, all the tests continue to compile with it removed.
Consequently, I think it makes sense to delete it.