Fix #1184 (`cats-effect` version conflict) by dropping Monix for now
This change updates to the latest version of `fs2-core`, which is now
compatible with `cats-effect` v3.1.1 (introduced in
https://github.com/scanamo/scanamo/pull/1166). Unfortunately there's not
yet a version of `monix` that supports `cats-effect` v3.1.1, so we're
dropping support for Monix for now - the actual implementation code
within Scanamo for the Monix support is only 4 lines, so can probably be
easily copied into an end-user project if it's needed!

Sbt is capable of warning us about version conflicts in dependencies, but
gave no warning of the `cats-effect` version conflict, I think because
the `fs2-core` & `monix` dependencies were marked as already `Provided`.
My guess as to why the `Provided` marker was used is that the
`scanamo-cats-effect` artifact gave support for both `fs2-core` & `monix`,
but in general a consuming project is likely to only be using one of
those - by declaring the two dependencies as `Provided`, neither will be
implicitly pulled in as a transitive dependency of `scanamo-cats-effect`
when used in the end-user project- giving the end-user-project freedom to
just explicitly declare a dependency on the one they want to use. As we've
dropped Monix support, and there's now only `fs2-core` support, I've
changed the `fs2-core` dependency to be explicit, which will give us the
benefit of sbt version-conflict warnings in future.

See https://github.com/scanamo/scanamo/issues/1184