Adopt GHA Scala Library Release Workflow
As mentioned in https://github.com/scanamo/scanamo/issues/1763, we're
switching to use `gha-scala-library-release-workflow` for releasing the
Scanamo to Maven Central. This is a change from the release process based
on `sbt-ci-release` that was introduced with https://github.com/scanamo/scanamo/pull/280 -
it used to be that to trigger a release, we would push a git tag with a manually-selected
version number (eg `v1.0.3`), the new process is now to run the `Release` workflow manually,
which will automatically determine a version number based on the binary-compatibility checks
performed by `sbt-version-policy`.

In order to do this, we've had to follow the one-off organisation-setup steps in:

https://github.com/guardian/gha-scala-library-release-workflow/blob/main/docs/org-setup.md

...which included creating a GitHub App for the Scanamo organisation:

https://github.com/apps/scala-library-release-scanamo

Additionally, this PR contains the per-repo modifications required by:

https://github.com/guardian/gha-scala-library-release-workflow/blob/main/docs/configuration.md