Remove runTest during release for now
Following on from https://github.com/scanamo/scanamo/pull/1769,
and as mentioned in:

https://github.com/guardian/gha-scala-library-release-workflow/blob/main/docs/configuration.md#recommended-sbt-settings

...unfortunately it's difficult to figure out how to modify
gha-scala-library-release-workflow to start the DynamodDB Local instance
to get tests to pass. So, either we modify Scanamo's sbt build to
automatically start the DynamodDB Local instance when the tests start -
or we just skip tests during release right now.