Improve transaction error-handling capabilities
This is based on the 2nd of the 2 distinct changes proposed in PR #1776,
which I'm now separating out into 2 PRs. Luciano Preissler's description
for _this_ change in https://github.com/scanamo/scanamo/pull/1776 was:

> Exposes TransactionCanceledException:
> This enhancement strengthens error handling capabilities, enabling users
> to distinguish between different transaction cancellation scenarios, thereby
> facilitating improved troubleshooting and error resolution.

The changes here are somewhat smaller than the original changes in PR #1776,
due to the refactoring in https://github.com/scanamo/scanamo/pull/1786.