Refactor: DRY ScanamoOps interpreters
The ScanamoOps interpreter for each framework (Future, Cats, Zio, Pekko, etc) has the
job of translating each Scanamo operation type (like `Get`, `Put`, etc) into actual work
done in the type container used by that framework (for ScanamoSync, that's just `Id[_]`, for
ScanamoAsync, that's `Future[_]`, for ZIO it's `IO[_]`, etc).

In Functional Programming terms, as described by the Cats guide to the Free Monad
(https://typelevel.org/cats/datatypes/freemonad.html#_4-write-a-compiler-for-your-program),
`ScanamoOpsA[A]` and its implementations (`Get`, `Put`, etc) are the Algebraic Data
Type (ADT), and the ScanamoOps interpreter is a compiler that compiles the DSL into
effective language that can actually be run. This is also known as the
'Free Monad + Interpreter pattern'.

Scanamo currently supports 5 different ScanamoOps interpreters:

* ScanamoAsyncInterpreter & ScanamoSyncInterpreter, first introduced as `future` & `id`
  in https://github.com/scanamo/scanamo/pull/11, and then given their current names in
  https://github.com/scanamo/scanamo/pull/406
* CatsInterpreter - https://github.com/scanamo/scanamo/pull/212
* ZioInterpreter - https://github.com/scanamo/scanamo/pull/262
* PekkoInterpreter - https://github.com/scanamo/scanamo/pull/1660, a port of
  AlpakkaInterpreter added with https://github.com/scanamo/scanamo/pull/151

Each interpreter has a `ScanamoOpsA[A]` match-case expression which is quite lengthy and
repetitive. Whenever a new ScanamoOp is added, or its return type modified (eg as in
https://github.com/scanamo/scanamo/pull/1776) there need to be changes to all 5
interpreters, and as naturally happens over time we get a lot of copy-n-paste,
to the point where the simplicity of the interpreter becomes hard to see. In the
case of #1776, which just follows the convention of the interpreter code already
in place, in commit 205a33f7 we see that in PekkoInterpreter, 1 line of code
becomes 5, and in CatsInterpreter 1 line becomes 10. Additionally, the types
get a bit verbose (eg `Either[TransactionCanceledException, TransactWriteItemsResponse]`).

Consequently, here I'm introducing a refactor that aims to address these issues:

* AsyncFrameworkInterpreter allows several interpreters (ScanamoAsyncInterpreter,
  CatsInterpreter, ZioInterpreter) to share a large portion of their implementation,
  including the large match-case expression
* helper methods `runConditional`, `runPaginated` (and later `runTransact`) remove a lot
  of noise from several cases in the match-case expression, especially `PekkoInterpreter`.
  There was already one method like these in `ZioInterpreter`, `effEitherConditionalCheckFailed`,
  but this wasn't applied more generally across the other interpreters
* Type aliases `Conditional` & `Transact` reduce noise from the type declarations
  that involve exposing exceptions with an Either, eg `Either[ConditionalCheckFailedException, _]`