Simplify CatsAdapter exception-handling
The original implementation of the code to expose the `ConditionalCheckFailedException`
was introduced with PR #212 in May 2018, but I'm not sure why the multi-line implementation
was necessary - there's a discussion in PR #212, but I don't quite follow it: https://github.com/scanamo/scanamo/pull/212#discussion_r187699312

The Cats `recover` functionality I'm using here was added with `ApplicativeError` in
https://github.com/typelevel/cats/pull/812 in January 2016, so it would have been available
for PR #212. Perhaps it was some consequence of us being on AWS SDK v1 at that point, and
having to use `AsyncHandler`?

This implementation is definitely much more similar to ones we've used elsewhere, ie
`ScalaFutureAdapter` and `PekkoInterpreter`.

If we had a `ClassTag` we could go even smaller and use `attemptNarrow`, which was added
to Cats with https://github.com/typelevel/cats/pull/2863 in May 2019 - but even without that,
the implementation is much smaller.