Extend AWS SDK builders to support `Option` setting
Scanamo's own Scala-based models of DynamoDB requests use Scala's idiomatic
approach to optional values - use `Option`, but this repeatedly gives us a
fiddly job when coping with the fluent nature of the AWS SDK for Java, which
chains methods on builders. We want to pretend that the builders aren't mutable
(because in Scala, they wouldn't be), and using the default methods on Option,
the most concise way to do that for a builder is using
`optionValue.fold(builder)(value => builder.setter(value)` - which isn't very
concise at all.

This change introduces a `setOpt` extension method on all AWS SDK `Buildable`s,
which provides a more concise optional-invocation of the builder's setter methods,
while supporting the same fluent chaining that native builder setter methods use.