Pekko: Enable Scala 3 support
Scala 3 support was added to Scanamo with
https://github.com/scanamo/scanamo/pull/1577 / https://github.com/scanamo/scanamo/pull/1596 in
November 2022, but as mentioned in that 2nd PR, we only enabled Scala 3 support for Scanamo
subprojects where it was nice & easy to do so - amongst others, that excluded Alpakka (the
predecessor to Pekko, originally added to Scanamo with https://github.com/scanamo/scanamo/pull/151
in October 2017).

When Pekko was added to Scanamo (porting the Alpakka support) with
https://github.com/scanamo/scanamo/pull/1660 in September 2023, it retained the constraint of
not supporting Scala 3 - we're fixing that here.

* Scala 3 cross-compilation enabled for the Pekko project
* Pekko tests enabled under the Scala 3 CI builds
* Fixed compilation error in `ScanamoPekko` relating to Scala 3's syntax for wildcard arguments
  in types - we can switch to the `Pekko[A]` type alias for `Source[A, NotUsed]` and avoid specifying
  the unwanted type parameters anyhow. See https://github.com/scanamo/scanamo/pull/1805/files#r1731496767
* Fix compilation error in `ScanamoPekkoSpec` due to changed implicit precedence in Scala 3
  which led to the wrong `===` method being selected - using one from `TripleEqualsSupport` in
  Scalactic rather than our own Scanamo one. See https://github.com/scanamo/scanamo/pull/1805/files#r1731491275